Rename command to configure
